Refactor: split and organize usercfg package
